requirements.txt
fritzconnection
pycurl
python-dotenv
